extract idempotency support bundle route
